UI-38 rename files
