+++ src/index.ts
@@ -14,6 +14,7 @@ export * from './BasicList'
 // import { List } from './List'
 export * from './Button'
 export * from './TextInput'
+export * from './RadioGroup'
 export * from './SearchInput'
 export * from './Footer'
 export * from './DateInput'
